requirements.txt
Django
Pillow
django-crispy-forms
crispy-bootstrap4
django-markdownx